refactor: event바인딩 함수 index.js에서 분리
