+++ js/index.js
@@ -1,26 +1,8 @@
 import { $ } from "./utils/dom.js";
-import { toggleClass, toggleWeekDayMenu, preventDefaults } from "./tools.js";
+import { toggleWeekDayMenu, preventDefaults } from "./tools.js";
+import { bindSubMenuEvent } from "./submenu.js";
 import { bindCaroulselEvent } from "./carousel.js";
-import { render, renderHome, renderWeekday } from "./renderer.js";
-
-const bindSubMenuEvent = () => {
-    $(".sub-menu").addEventListener("click", ({ target }) => {
-        const curEl = target.parentNode;
-        const targetPage = target.innerText;
-
-        toggleClass(curEl, "active");
-
-        if (targetPage === "홈") {
-            renderHome();
-            return;
-        }
-        if (targetPage === "요일연재") {
-            renderWeekday();
-            return;
-        }
-        render("");
-    });
-};
+import { renderHome, renderWeekday } from "./renderer.js";
 
 const bindEventListeners = () => {
     bindSubMenuEvent();
